feat(*): Adds wash plugin for wit2wadm
Adds a wash plugin version for wit2wadm. I tested this with wash 0.28
to ensure that everything works. Please note that I had to copy some
code from wit-parser to make sure it could properly parse a wit dir so
there is quite a bit of code there. This also refactors some things for
reusability in the plugin and the main binary.

As a follow up, we should probably add a workflow for publishing this
as an OCI artifact

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>